Really just a guessing game
I wrote this over a year ago to show to my friends. initially made to be a code breaker but...ended up being a guessing game. Still, something that's nice to look back on! I wish i put loops so that I wouldn't have to run the program every time something goes wrong.