From March 28, 2023
This had to do with my research into God and some other things. It calculates how long it would take to roll a dice or flip a coin and get teh same result if each reroll takes x seconds. Not sure if it works too too well considering some things like saying a dice has 6 sides but you want it to land on the number 8.